+++ packages/skeleton-react/src/lib/utils/react-compose.ts
@@ -1,0 +1,7 @@
+export function reactCompose<
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
+	T extends React.FC<any>,
+	U extends Record<string, unknown>
+>(root: T, components: { [k in keyof U]: React.FC<U[k]> }) {
+	return Object.assign(root, components);
+}
