--- packages/skeleton-react/src/utils/react-compose.ts
@@ -1,6 +1,0 @@
-export function reactCompose<
-  T extends React.FC<any>,
-  U extends Record<string, unknown>
->(root: T, components: { [k in keyof U]: React.FC<U[k]> }) {
-  return Object.assign(root, components);
-}
